requirements.txt
tensorflow>=2.10.0
transformers>=4.25.1
nltk>=3.7
pandas>=1.5.2
numpy>=1.23.5
scikit-learn>=1.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
matplotlib>=3.6.2
streamlit>=1.16.0
